refactoring hook
