+++ assets/resources/answer.json
@@ -1,0 +1,182 @@
+{
+  "1": [3, 1, 9],
+  "2": [2, 0, 8],
+  "3": [7, 5, 3],
+  "4": [6, 4, 2],
+  "5": [9, 7, 5],
+  "6": [4, 2, 0],
+  "7": [8, 6, 4],
+  "8": [0, 8, 6],
+  "9": [1, 9, 7],
+  "10": [5, 3, 1],
+  "11": [6, 8, 7],
+  "12": [8, 0, 9],
+  "13": [9, 1, 0],
+  "14": [4, 6, 5],
+  "15": [3, 5, 4],
+  "16": [7, 9, 8],
+  "17": [2, 4, 3],
+  "18": [5, 7, 6],
+  "19": [1, 3, 2],
+  "20": [0, 2, 1],
+  "21": [3, 2, 4],
+  "22": [9, 8, 0],
+  "23": [2, 1, 3],
+  "24": [1, 0, 2],
+  "25": [4, 3, 5],
+  "26": [0, 9, 1],
+  "27": [6, 5, 7],
+  "28": [7, 6, 8],
+  "29": [8, 7, 9],
+  "30": [5, 4, 6],
+  "31": [6, 2, 5],
+  "32": [3, 9, 2],
+  "33": [0, 6, 9],
+  "34": [9, 5, 8],
+  "35": [7, 3, 6],
+  "36": [5, 1, 4],
+  "37": [2, 8, 1],
+  "38": [8, 4, 7],
+  "39": [1, 7, 0],
+  "40": [4, 0, 3],
+  "41": [0, 4, 6],
+  "42": [6, 0, 2],
+  "43": [4, 8, 0],
+  "44": [8, 2, 4],
+  "45": [5, 9, 1],
+  "46": [9, 3, 5],
+  "47": [2, 6, 8],
+  "48": [7, 1, 3],
+  "49": [3, 7, 9],
+  "50": [1, 5, 7],
+  "51": [0, 6, 4],
+  "52": [6, 2, 0],
+  "53": [8, 4, 2],
+  "54": [3, 9, 7],
+  "55": [4, 0, 8],
+  "56": [2, 8, 6],
+  "57": [1, 7, 5],
+  "58": [9, 5, 3],
+  "59": [7, 3, 1],
+  "60": [5, 1, 9],
+  "61": [7, 6, 3],
+  "62": [2, 1, 8],
+  "63": [4, 3, 0],
+  "64": [1, 0, 7],
+  "65": [6, 5, 2],
+  "66": [9, 8, 5],
+  "67": [5, 4, 1],
+  "68": [8, 7, 4],
+  "69": [3, 2, 9],
+  "70": [0, 9, 6],
+  "71": [7, 3, 8],
+  "72": [2, 8, 3],
+  "73": [5, 1, 6],
+  "74": [4, 0, 5],
+  "75": [8, 4, 9],
+  "76": [0, 6, 1],
+  "77": [6, 2, 7],
+  "78": [9, 5, 0],
+  "79": [3, 9, 4],
+  "80": [1, 7, 2],
+  "81": [1, 0, 3],
+  "82": [0, 9, 2],
+  "83": [7, 6, 9],
+  "84": [5, 4, 7],
+  "85": [9, 8, 1],
+  "86": [6, 5, 8],
+  "87": [2, 1, 4],
+  "88": [4, 3, 6],
+  "89": [3, 2, 5],
+  "90": [8, 7, 0],
+  "91": [0, 4, 9],
+  "92": [1, 5, 0],
+  "93": [5, 9, 4],
+  "94": [4, 8, 3],
+  "95": [9, 3, 8],
+  "96": [6, 0, 5],
+  "97": [2, 6, 1],
+  "98": [3, 7, 2],
+  "99": [8, 2, 7],
+  "100": [7, 1, 6],
+  "101": [9, 1, 7],
+  "102": [3, 5, 1],
+  "103": [6, 8, 4],
+  "104": [2, 4, 0],
+  "105": [8, 0, 6],
+  "106": [5, 7, 3],
+  "107": [1, 3, 9],
+  "108": [7, 9, 5],
+  "109": [0, 2, 8],
+  "110": [4, 6, 2],
+  "111": [5, 8, 4],
+  "112": [3, 6, 2],
+  "113": [7, 0, 6],
+  "114": [0, 3, 9],
+  "115": [1, 4, 0],
+  "116": [4, 7, 3],
+  "117": [2, 5, 1],
+  "118": [8, 1, 7],
+  "119": [6, 9, 5],
+  "120": [9, 2, 8],
+  "121": [4, 1, 8],
+  "122": [5, 2, 9],
+  "123": [7, 4, 1],
+  "124": [1, 8, 5],
+  "125": [8, 5, 2],
+  "126": [6, 3, 0],
+  "127": [3, 0, 7],
+  "128": [2, 9, 6],
+  "129": [0, 7, 4],
+  "130": [9, 6, 3],
+  "131": [2, 9, 7],
+  "132": [5, 2, 0],
+  "133": [3, 0, 8],
+  "134": [4, 1, 9],
+  "135": [9, 6, 4],
+  "136": [0, 7, 5],
+  "137": [1, 8, 6],
+  "138": [7, 4, 2],
+  "139": [6, 3, 1],
+  "140": [8, 5, 3],
+  "141": [3, 5, 9],
+  "142": [1, 3, 7],
+  "143": [5, 7, 1],
+  "144": [2, 4, 8],
+  "145": [0, 2, 6],
+  "146": [8, 0, 4],
+  "147": [7, 9, 3],
+  "148": [4, 6, 0],
+  "149": [9, 1, 5],
+  "150": [6, 8, 2],
+  "151": [8, 1, 4],
+  "152": [9, 2, 5],
+  "153": [6, 9, 2],
+  "154": [2, 5, 8],
+  "155": [0, 3, 6],
+  "156": [1, 4, 7],
+  "157": [3, 6, 9],
+  "158": [4, 7, 0],
+  "159": [7, 0, 3],
+  "160": [5, 8, 1],
+  "161": [7, 1, 8],
+  "162": [1, 5, 2],
+  "163": [2, 6, 3],
+  "164": [5, 9, 6],
+  "165": [3, 7, 4],
+  "166": [0, 4, 1],
+  "167": [4, 8, 5],
+  "168": [8, 2, 9],
+  "169": [9, 3, 0],
+  "170": [6, 0, 7],
+  "171": [8, 4, 1],
+  "172": [0, 6, 3],
+  "173": [3, 9, 6],
+  "174": [4, 0, 7],
+  "175": [6, 2, 9],
+  "176": [9, 5, 2],
+  "177": [1, 7, 4],
+  "178": [5, 1, 8],
+  "179": [7, 3, 0],
+  "180": [2, 8, 5]
+}
