+++ src/renderer/index.tsx
@@ -11,13 +11,14 @@ window.electron.ipcRenderer.once('ipc', (arg) => {
   console.log(arg);
 });
 
-log.info('Log from the renderer process');
-
 window.electron.ipcRenderer.sendMessage('ipc', ['ping']);
+
 window.electron.ipcRenderer.on('sql-file-failured', (message) => {
   alert(message);
 });
+
 window.electron.ipcRenderer.on('save-file-failured', (message) => {
   alert(message);
 });
 
+log.info('Log from the renderer process');
