+++ src/renderer/services/store/slices/answerProvider.ts
@@ -1,0 +1,14 @@
+import { createSlice } from '@reduxjs/toolkit';
+import answerData from '@assets/resources/answer.json';
+
+type dataType = typeof answerData;
+
+export const answerSlice = createSlice({
+  name: 'answers',
+  initialState: answerData,
+  reducers: {}
+});
+
+export const getAnswers = (key: string | number) => answerSlice.getInitialState()[key as keyof dataType];
+
+export default answerSlice.reducer;
